Split dual_axis.rs
